cleanup

+++ 2016/Day25.hs
@@ -14,7 +14,7 @@ So to yield an alternating string 0101..., a+2534 must be of the form
 -}
 
 solve1 :: Int
-solve1 = head $ dropWhile (<= 0) [(4^k-1)*2 `div` 3 - 2534 | k <- [1..]]
+solve1 = head $ dropWhile (<= 0) [(4^k-1)*2 `div` 3 - 2534 | k <- [(1::Int)..]]
 
 main :: IO ()
 main = print solve1
